+++ libsrc/ffdec_lib/src/com/jpexs/csv/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * CSV files parsing
+ */
+package com.jpexs.csv;
+